+++ Changelog.md
@@ -2,6 +2,10 @@
 
 ### Ignition Gazebo 2.X.X
 
+1. Level performers can be added at runtime using a service call. See the
+   levels tutorial for more information.
+    * [Pull Request 107](https://bitbucket.org/ignitionrobotics/ign-gazebo/pull-requests/107)
+
 1. Update PosePublisher system to publish sensor poses and to use scoped names for frame ids
     * [Pull Request 331](https://bitbucket.org/ignitionrobotics/ign-gazebo/pull-requests/331)